fix build of COOL Lite
