+++ wsd/ClientRequestDispatcher.cpp
@@ -36,6 +36,9 @@
 #if !MOBILEAPP
 #include <HostUtil.hpp>
 #endif // !MOBILEAPP
+#if ENABLE_SUPPORT_KEY
+#include <UserMessages.hpp>
+#endif
 
 #include <Poco/DOM/AutoPtr.h>
 #include <Poco/DOM/DOMParser.h>
@@ -64,6 +67,35 @@ extern std::map<std::string, std::shared_ptr<DocumentBroker>> DocBrokers;
 extern std::mutex DocBrokersMutex;
 
 extern void cleanupDocBrokers();
+
+namespace
+{
+
+#if ENABLE_SUPPORT_KEY
+inline void shutdownLimitReached(const std::shared_ptr<ProtocolHandlerInterface>& proto)
+{
+    if (!proto)
+        return;
+
+    const std::string error = Poco::format(PAYLOAD_UNAVAILABLE_LIMIT_REACHED, COOLWSD::MaxDocuments, COOLWSD::MaxConnections);
+    LOG_INF("Sending client 'hardlimitreached' message: " << error);
+
+    try
+    {
+        // Let the client know we are shutting down.
+        proto->sendTextMessage(error.data(), error.size());
+
+        // Shutdown.
+        proto->shutdown(true, error);
+    }
+    catch (const std::exception& ex)
+    {
+        LOG_ERR("Error while shutting down socket on reaching limit: " << ex.what());
+    }
+}
+#endif
+
+} // end anonymous namespace
 
 /// Find the DocumentBroker for the given docKey, if one exists.
 /// Otherwise, creates and adds a new one to DocBrokers.
